+++ Core.lua
@@ -427,10 +427,4 @@ addon.config.onLoad = function(self)
     -- Settings subcategory
     BazCore:RegisterOptionsTable(ADDON_NAME .. "-Settings", GetSettingsPage)
     BazCore:AddToSettings(ADDON_NAME .. "-Settings", "Settings", ADDON_NAME)
-
-    -- Profiles
-    BazCore:RegisterOptionsTable(ADDON_NAME .. "-Profiles", function()
-        return BazCore:GetProfileOptionsTable(ADDON_NAME)
-    end)
-    BazCore:AddToSettings(ADDON_NAME .. "-Profiles", "Profiles", ADDON_NAME)
 end